Pause CADisplayLink during when backgrounded
